backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart

langchain
langchain-community
langchain-core

transformers
sentence-transformers
huggingface-hub
accelerate
torch

# Vectorstore
# chromadb
pinecone

# Embeddings
# sentence-transformers
# langchain-google-generative-ai
langchain-google-genai

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests
tqdm

# Logging (optional but recommended)
loguru
